break out configurable. add taps. start checking names/labels with regex
