+++ src/CMakeLists.txt
@@ -11,7 +11,8 @@ add_library(visor-core
         HandlerModulePlugin.cpp
         GeoDB.cpp
         CoreServer.cpp
-        Metrics.cpp Metrics.h)
+        Metrics.cpp
+        Taps.cpp)
 add_library(Visor::Core ALIAS visor-core)
 
 target_include_directories(visor-core
