+++ src/Configurable.h
@@ -1,0 +1,81 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+#pragma once
+
+#include <atomic>
+#include <exception>
+#include <nlohmann/json.hpp>
+#include <shared_mutex>
+#include <string>
+#include <unordered_map>
+#include <variant>
+
+namespace visor {
+
+using json = nlohmann::json;
+
+class ConfigException : public std::runtime_error
+{
+public:
+    explicit ConfigException(const std::string &msg)
+        : std::runtime_error(msg)
+    {
+    }
+};
+
+class Configurable
+{
+private:
+    std::unordered_map<std::string, std::variant<std::string, uint64_t, bool>> _config;
+    mutable std::shared_mutex _config_mutex;
+
+public:
+    template <class T>
+    auto config_get(const std::string &key)
+    {
+        std::shared_lock lock(_config_mutex);
+        if (_config.count(key) == 0) {
+            throw ConfigException("missing key: " + key);
+        }
+        auto val = std::get_if<T>(&_config[key]);
+        if (!val) {
+            throw ConfigException("wrong type for key: " + key);
+        }
+        return *val;
+    }
+
+    template <class T>
+    void config_set(const std::string &key, const T &val)
+    {
+        std::unique_lock lock(_config_mutex);
+        _config[key] = val;
+    }
+
+    // specialize to ensure a string literal is interpreted as a std::string
+    void config_set(const std::string &key, const char *val)
+    {
+        std::unique_lock lock(_config_mutex);
+        _config[key] = std::string(val);
+    }
+
+    bool config_exists(const std::string &name) const
+    {
+        std::shared_lock lock(_config_mutex);
+        return _config.count(name) == 1;
+    }
+
+    void config_json(json &j) const
+    {
+        std::shared_lock lock(_config_mutex);
+        for (const auto &[key, value] : _config) {
+            std::visit([&j, key = key](auto &&arg) {
+                j[key] = arg;
+            },
+                value);
+        }
+    }
+};
+
+}
